Error checking & Clean-up Fixed a bug with 0 sized or negative integer maps causing segmentation faults, added error checking to mallocs, map sizes and (sanitized?) the free_level() function. (Credits to 0xcaca0)

+++ src/soko_sdl.c
@@ -47,7 +47,7 @@ void _render(SDL_Surface *screen, SDL_Surface **assets, int asset_width, int ass
         for (j = 0; j < map->height; j++) {
             tmp.x = i * asset_width;
             tmp.y = j * asset_height;
-            SDL_BlitSurface(assets[map->data[j][i]], NULL, screen, &tmp);
+            SDL_BlitSurface(assets[(int)map->data[j][i]], NULL, screen, &tmp);
         }
     }
